chore: 新增 npm 命令 dev:web，支持仅启动 web 版本的开发环境

+++ .erb/configs/webpack.config.renderer.dev.ts
@@ -24,6 +24,8 @@ const skipDLLs =
     module.parent?.filename.includes('webpack.config.renderer.dev.dll') ||
     module.parent?.filename.includes('webpack.config.eslint')
 
+const DEV_WEB_ONLY = process.env.DEV_WEB_ONLY === 'true'
+
 /**
  * Warn if the DLL is not built
  */
@@ -116,12 +118,12 @@ const configuration: webpack.Configuration = {
         ...(skipDLLs
             ? []
             : [
-                  new webpack.DllReferencePlugin({
-                      context: webpackPaths.dllPath,
-                      manifest: require(manifest),
-                      sourceType: 'var',
-                  }),
-              ]),
+                new webpack.DllReferencePlugin({
+                    context: webpackPaths.dllPath,
+                    manifest: require(manifest),
+                    sourceType: 'var',
+                }),
+            ]),
 
         new webpack.NoEmitOnErrorsPlugin(),
 
@@ -193,20 +195,22 @@ const configuration: webpack.Configuration = {
                 .on('close', (code: number) => process.exit(code!))
                 .on('error', (spawnError) => console.error(spawnError))
 
-            console.log('Starting Main Process...')
-            let args = ['run', 'start:main']
-            if (process.env.MAIN_ARGS) {
-                args = args.concat(['--', ...process.env.MAIN_ARGS.matchAll(/"[^"]+"|[^\s"]+/g)].flat())
-            }
-            spawn('npm', args, {
-                shell: true,
-                stdio: 'inherit',
-            })
-                .on('close', (code: number) => {
-                    preloadProcess.kill()
-                    process.exit(code!)
+            if (!DEV_WEB_ONLY) {
+                console.log('Starting Main Process...')
+                let args = ['run', 'start:main']
+                if (process.env.MAIN_ARGS) {
+                    args = args.concat(['--', ...process.env.MAIN_ARGS.matchAll(/"[^"]+"|[^\s"]+/g)].flat())
+                }
+                spawn('npm', args, {
+                    shell: true,
+                    stdio: 'inherit',
                 })
-                .on('error', (spawnError) => console.error(spawnError))
+                    .on('close', (code: number) => {
+                        preloadProcess.kill()
+                        process.exit(code!)
+                    })
+                    .on('error', (spawnError) => console.error(spawnError))
+            }
             return middlewares
         },
     },
